% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_node_installed}
\alias{check_node_installed}
\title{Check if node is installed}
\usage{
check_node_installed()
}
\description{
Put this function in \link{.onAttach} for in your package containing a node
module.  It will stop package loading (by throwing an error) if node is not
available, and print a message if it is.
}

