% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dat.R
\name{dat}
\alias{dat}
\title{Dat repository}
\usage{
dat(path = tempdir(), dat = "dat-beta", verbose = TRUE)
}
\arguments{
\item{path}{directory of the dat repository}

\item{dat}{name of the 'dat' executable  (possibly with path)}

\item{verbose}{gives some more output}
}
\description{
Create and modify a dat repository.
}
\examples{
repo <- dat(tempdir())
repo$import(mtcars)
repo$export()
}

