% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dat_init}
\alias{dat_init}
\title{Initialize a dat repository}
\usage{
dat_init(path = ".")
}
\arguments{
\item{path}{Path to data store. use \code{"."} for current dir.}
}
\description{
This function will initialize a dat repository
}
\examples{
\dontrun{
dat_init('.')
}
}
\keyword{initialize}

