% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dat.R
\name{dat}
\alias{dat}
\title{Dat repository}
\usage{
dat(dataset = "test", remote = NULL, path = tempdir(), dat = "dat",
  verbose = FALSE)
}
\arguments{
\item{dataset}{name of the dat 'dataset' (namespace)}

\item{path}{directory of the dat repository}

\item{dat}{name of the 'dat' executable  (possibly with path)}

\item{verbose}{gives some more output}
}
\description{
Create and modify a dat repository.
}
\examples{
# init a temporary repo
repo <- dat("cars")

# insert some data
repo$insert(cars[1:20,])
v1 <- repo$status()$version
v1

# insert some more data
repo$insert(cars[21:25,])
v2 <- repo$status()$version
v2

# get the data
data1 <- repo$get(v1)
data2 <- repo$get(v2)
diff <- repo$diff(v1, v2)
diff$key

# create fork
repo$checkout(v1)
repo$insert(cars[26:30,])
repo$forks()
v3 <- repo$status()$version

# go back
repo$checkout(v2)
repo$get()

# store binary attachements
repo$write(serialize(iris, NULL), "iris")
unserialize(repo$read("iris"))

# Create another repo
dir.create(newdir <- tempfile())
repo2 <- dat("cars", path = newdir, remote = repo$path())
repo2$forks()
repo2$get()
repo2$insert(cars[31:40,])

# Create a third repo
dir.create(newdir <- tempfile())
repo3 <- dat("cars", path = newdir, remote = repo$path())

# Sync 2 with 3 via remote (1)
repo2$push()
repo3$pull()
mydata2 <- repo2$get()
mydata3 <- repo3$get()
all.equal(mydata2, mydata3)
}

